rename: isNotOverThanMax -> isNotOverMaxLength로 함수 이름 변경
