+++ hooks/src/lib/useSingleNumberInput.ts
@@ -1,12 +1,12 @@
 import { useState } from 'react';
-import { isNotOverThanMax, isNumeric } from './utils';
+import { isNotOverMaxLength, isNumeric } from './utils';
 
 const useSingleNumberInput = (maxLength: number) => {
   const [inputValue, setInputValue] = useState('');
 
   const errorText = (() => {
     if (!isNumeric(inputValue)) return '입력값이 숫자가 아닙니다.';
-    if (!isNotOverThanMax(inputValue, maxLength)) return '입력값이 최대 길이를 초과했습니다.';
+    if (!isNotOverMaxLength(inputValue, maxLength)) return '입력값이 최대 길이를 초과했습니다.';
     return '';
   })();
