+++ hooks/src/lib/utils/index.ts
@@ -1,2 +1,2 @@
 export { default as isNumeric } from './isNumeric';
-export { default as isNotOverThanMax } from './isNotOverThanMax';
+export { default as isNotOverMaxLength } from './isNotOverMaxLength';
