+++ hooks/src/lib/utils/isNotOverMaxLength.ts
@@ -1,0 +1,5 @@
+const isNotOverMaxLength = (value: string, maxLength: number) => {
+  return value.length <= maxLength;
+};
+
+export default isNotOverMaxLength;
