--- hooks/src/lib/utils/isNotOverThanMax.ts
@@ -1,5 +1,0 @@
-const isNotOverThanMax = (value: string, maxLength: number) => {
-  return value.length <= maxLength;
-};
-
-export default isNotOverThanMax;
